storytelling/requirements.txt
google-adk[extensions]~=1.26.0
fastapi
litellm
uvicorn
